{{ cookiecutter.project_slug }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_name }}"
version = "0.1.0"
description = "{{ cookiecutter.project_description }}"
authors = ["Elation Health, Inc. <no-reply@elationhealth.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "el8"
url = "https://el8-global-570488747013.d.codeartifact.us-west-2.amazonaws.com/pypi/el8-global-pypi/simple/"
priority = "default"

[tool.poetry.dependencies]
python = "~=3.10"
ddtrace = "~=1.20"
el8-py-ext-fastapi = "~=2.0"
pydantic = "~=2.5"

[tool.poetry.group.dev.dependencies]
poethepoet = "~=0.24"
pytest = "~=7.4"
pytest-cov = "~=4.1"
pytest-asyncio = "~=0.21"
black = "~=23.11"
mypy = "~=1.7"
flake8 = "~=6.1"
flake8-black = "~=0.3"
isort = "~=5.12"
boto3 = "~=1.29"
httpx = "~=0.25"
types-protobuf = "*"
types-setuptools = "*"
types-six = "*"
types-ujson = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe]
verbosity = -1

# https://poethepoet.natn.io/index.html
[tool.poe.tasks]
# define custom tasks here